DONE Cleanup code (Terminate plugin, etc.)
