+++ enki/plugins/pluginmanager/helper.py
@@ -1,0 +1,70 @@
+"""Helper functions for the pluginmanager plugin"""
+
+import os
+import shutil
+from enki.core.core import core
+
+from .constants import *
+
+# Data Definitions
+# ==================
+# userpluginEntry consist of
+# {'module': Module
+#  'plugin': Plugin or None,
+#  'isloaded': Bool,
+#  'name': String,
+#  'author': String,
+#  'version': String,
+#  'doc': String}
+#
+# ListOfUserpluginEntry is one of:
+# - []
+# - lou.append(userpluginEntry)
+
+def create_UE(module, isLoaded, modulename, pluginname,
+    author, version, doc, plugin=None):
+    """Create a new userpluginEntry"""
+    return {'module': module,
+            'plugin': plugin,
+            'isLoaded': isLoaded,
+            'modulename': modulename,
+            'pluginname': pluginname,
+            'author': author,
+            'version': version,
+            'doc': doc}
+
+def loadPlugin(pluginEntry):
+    """Consume a pluginEntry and load the plugin into core._loadedPlugins,
+    based on it's 'isLoaded' property
+    return pluginEntry
+    """
+    if pluginEntry['isLoaded'] == True:
+        pluginEntry['plugin'] = pluginEntry['module'].Plugin()
+        core.loadedPlugins().append(pluginEntry['plugin'])
+    return pluginEntry
+
+def unloadPlugin(pluginEntry):
+    """Consume a pluginEntry and unload the plugin from core._loadedPlugins,
+    based on it's 'isLoaded' property
+    return pluginEntry
+    """
+    if pluginEntry['isLoaded'] is False and \
+        pluginEntry['plugin'] is not None:
+        pluginEntry['plugin'].terminate()
+        idx = core.loadedPlugins().index(pluginEntry['plugin'])
+        core.loadedPlugins().pop(idx)
+        pluginEntry['plugin'] = None
+    return pluginEntry
+
+def deletePlugin(pluginEntry):
+    """Consume a pluginEntry and delete the plugin directory or file"""
+    unloadPlugin(pluginEntry)
+    dirpath = os.path.join(_PLUGIN_DIR_PATH, pluginEntry["modulename"])
+    filepath = os.path.join(_PLUGIN_DIR_PATH, pluginEntry["modulename"] + '.py')
+    if  os.path.isdir(dirpath):
+        shutil.rmtree(dirpath)
+    elif os.path.exists(filepath):
+        os.remove(filepath)
+    else:
+        print("Could not find module %s. Did not delete anything." %
+            pluginEntry["modulename"])
